# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      responses:
        "200":
          description: Tenants the current user belongs to
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tenant" }
    post:
      operationId: createTenant
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantInput" }
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
  /tenants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getTenant
      tags: [tenants]
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
    patch:
      operationId: updateTenant
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantUpdate" }
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
    delete:
      operationId: deleteTenant
      tags: [tenants]
      responses:
        "204":
          description: Deleted
  /tenants/{tenantId}/memberships:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listMemberships
      tags: [tenants]
      responses:
        "200":
          description: Memberships for tenant
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Membership" }
    post:
      operationId: createMembership
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MembershipInput" }
      responses:
        "201":
          description: Created membership
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
  /memberships/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateMembership
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MembershipUpdate" }
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Membership" }
    delete:
      operationId: deleteMembership
      tags: [tenants]
      responses:
        "204":
          description: Deleted

  # ---------- Hubs / Domains (global read-only reference data) ----------
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, name, slug, plan]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [free, starter, pro, enterprise] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TenantInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        plan: { type: string, enum: [free, starter, pro, enterprise] }
    TenantUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [free, starter, pro, enterprise] }
    Membership:
      type: object
      required: [id, tenantId, userId, role, status]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        userId: { type: string }
        role: { type: string, enum: [owner, admin, member, viewer] }
        status: { type: string, enum: [pending, active, revoked] }
        invitedBy: { type: ["string", "null"] }
        invitedEmail: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MembershipInput:
      type: object
      required: [userId]
      properties:
        userId: { type: string }
        role: { type: string, enum: [owner, admin, member, viewer] }
        status: { type: string, enum: [pending, active, revoked] }
        invitedEmail: { type: string }
    MembershipUpdate:
      type: object
      properties:
        role: { type: string, enum: [owner, admin, member, viewer] }
        status: { type: string, enum: [pending, active, revoked] }

    # ---------- Hubs ----------
...