# lib/api-zod/src/generated/types/membershipUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipUpdateRole } from './membershipUpdateRole';
import type { MembershipUpdateStatus } from './membershipUpdateStatus';

export interface MembershipUpdate {
  role?: MembershipUpdateRole;
  status?: MembershipUpdateStatus;
}